bug fixes
#168

This issue effected both questions that had more than one selectable checkbox.  When selecting a checkbox, it enables the Continue button, however if you selected more than one, and then unselect a checkbox, it will disable the Continue button.  This now checks all of the checkboxes to see if any is still selected, and only disables the Continue button if none are checked.